Use same structure for HAR and DOM advice results
When we changed the DOM structure to work with Browsertime we broke
the merging of DOM & HAR. We din't had good tests up and running
so we missed it.

HAR now follows the same structure as DOM and we have a simple test
that can be better in the future. #42